Fixed a bug in the tile set up.
The white rook on the eighth rank was set up in the back end (tile
array) as a black rook, which is why it had previously seemed immovable.
I figured this out when I tried moving the rook on black's turn, and saw
that the rook turned black when placed on another tile (meaning it's
color value must have been wrong in the tile array)

Next I will attempt to set up a text box that will display which color
is to play, as well as error messages when players attempt to make
illegal moves. Eventually, checks/checkmates will result in a message
here as well.