Fixed bugs in getKingMoves and reset methods.
Found a bug that prevented kings from moving on the 8th rank, and
realized that reset did not function properly.

Have been considering how to get a program to calculate the best move,
but am struggling to come up with a solution right now. For now, I will
enable en passant.