Fixed bug involving with isCheckMate
Bug actually was in getAllLegalMoves() and isLegalMove(), as they do not
take a tile array as an input, and therefore use the current active
tiles instead of the tiles as they would be after a given set of moves.

This bug only mattered in the isCheckMate method, as it is the only
method involved  with the 'ai' to call these methods.