'ai' improvements
I have attempted to implement preferences into the 'ai' by setting any
moves that are towards the center of the board, castles, or are pawns
moving on the opposite side of the board to have a higher weighted value
than others. I am not sure if it is working.

I will check on that tomorrow, and then I plan on adding two more layers
to the minimaxing algorithm to make the 'ai' much better, though I don't
feel confident about how fast it will be able to produce moves.