Bug fix/messed with AI
I tried to add two extra moves to the ai's simulation, but it took too
long. One more move also took too long (over 2 minutes per move).

I reverted the program back to the position it was in before the
changes, besides that I had fixed a couple of bugs.

Bugs still exist, as I just had a game tell me that white had no more
moves when it was not a stalemate or a checkmate. Also, black had had a
checkmate in 1 the move before and did not do it. :(

I will debug the program, and that will be the last update to the
program.