Worked on getMove method
Added helper methods to getMove to make code easier to digest. Made
ValueList (basically an ArrayList with two extra add methods) to help
reduce the time it takes to get a move (values has to be parsed through
one less time).

I will need to debug this, as the CPU was being an absolute moron when I
tested it. (likely swapped a "min" or "max" somewhere, shouldn't be too
tough)