Debugged GetWhite/BlackValues methods
Fixed an issue in which the index didn't update after a
checkmate/stalemate was reached prior to the third move.

The ai seems to be able to get a check mate, but for some reason cannot
defend against a potential checkmate in one. This must be a bug, as
before realigning my code the ai could do this. Will check on this.