Debugged
"Fixed" preference settings, as each time it got the move to check if it
was into the middle, was a king, etc. it was using the wrong index.
Still not sure if it works correctly, I may mess around with this more.

Also, got rid of a bug that caused a NullPointerException when black
castled short-side.